Adds style to hero section
Margin and padding set to 0 - resets the browsers default margin and padding for the entire html file.
Box-sizing to border-box - guarantees that the element size includes borders and padding.

CSS scroll snapping (.container(parent) and section(child))- allows full page scroll effect. Scroll snapping refers to "locking" the position of the viewport to specific elements on the page as the window (or a scrollable container) is scrolled.